+++ app/views/EditAry/AssessmentRegistryForm/SummaryForm/DimensionItem/SubDimensionItem/SelectIssueInput/index.tsx
@@ -1,5 +1,5 @@
-import React, { useCallback, useState, useMemo } from 'react';
-import { SearchSelectInput, TextInput } from '@the-deep/deep-ui';
+import React, { useCallback } from 'react';
+import { TextInput } from '@the-deep/deep-ui';
 import {
     EntriesAsList,
     useFormObject,
@@ -8,9 +8,7 @@ import {
     getErrorObject,
 } from '@togglecorp/toggle-form';
 import { randomString, isDefined, isNotDefined } from '@togglecorp/fujs';
-import { gql, useQuery } from '@apollo/client';
 
-import useDebouncedValue from '#hooks/useDebouncedValue';
 import {
     PartialFormType,
     SummaryIssueType,
@@ -19,34 +17,10 @@ import {
 import {
     AssessmentRegistrySectorTypeEnum,
     AssessmentRegistrySummarySubDimensionTypeEnum,
-    SummarySubDimensionIssueSearchQuery,
-    SummarySubDimensionIssueSearchQueryVariables,
 } from '#generated/types';
 
+import IssueSearchSelectInput from '../../../IssueSearchSelectInput';
 import styles from './styles.css';
-
-const SUMMARY_SUB_DIMENSION_ISSUE_SEARCH = gql`
-    query SummarySubDimensionIssueSearch(
-        $subDimension: AssessmentRegistrySummarySubDimensionTypeEnum,
-        $search: String,
-        $page: Int,
-        $pageSize: Int,
-    ) {
-        assessmentRegSummaryIssues(
-            subDimension: $subDimension,
-            search: $search,
-            pageSize: $pageSize,
-            page: $page,
-        ) {
-            page
-            results {
-                id
-                label
-            }
-            totalCount
-        }
-    }
-`;
 
 interface Props {
     name: string;
@@ -71,9 +45,6 @@ interface Props {
     sector: AssessmentRegistrySectorTypeEnum;
 }
 
-const keySelector = (d: SummaryIssueType) => d.id;
-const labelSelector = (d: SummaryIssueType) => d.label;
-
 function SelectIssueInput(props: Props) {
     const {
         subDimension,
@@ -90,10 +61,6 @@ function SelectIssueInput(props: Props) {
         error,
         disabled,
     } = props;
-
-    const [opened, setOpened] = useState(false);
-    const [searchText, setSearchText] = useState('');
-    const debouncedSearchText = useDebouncedValue(searchText);
 
     const {
         setValue: setSubDimensionIssue,
@@ -171,69 +138,6 @@ function SelectIssueInput(props: Props) {
         onChange,
     ]);
 
-    const variables = useMemo(
-        (): SummarySubDimensionIssueSearchQueryVariables => ({
-            subDimension: subDimension as AssessmentRegistrySummarySubDimensionTypeEnum,
-            search: debouncedSearchText,
-            page: 1,
-            pageSize: 10,
-        }), [
-            debouncedSearchText,
-            subDimension,
-        ],
-    );
-
-    const {
-        loading,
-        data,
-        fetchMore,
-    } = useQuery<
-        SummarySubDimensionIssueSearchQuery,
-        SummarySubDimensionIssueSearchQueryVariables
-    >(
-        SUMMARY_SUB_DIMENSION_ISSUE_SEARCH,
-        {
-            skip: !opened,
-            variables,
-        },
-    );
-
-    const handleShowMoreClick = useCallback(() => {
-        fetchMore({
-            variables: {
-                ...variables,
-                page: (data?.assessmentRegSummaryIssues?.page ?? 1) + 1,
-            },
-            updateQuery: (previousResult, { fetchMoreResult }) => {
-                if (!previousResult.assessmentRegSummaryIssues) {
-                    return previousResult;
-                }
-
-                const oldIssues = previousResult.assessmentRegSummaryIssues;
-                const newIssues = fetchMoreResult?.assessmentRegSummaryIssues;
-
-                if (!newIssues) {
-                    return previousResult;
-                }
-
-                return ({
-                    ...previousResult,
-                    assessmentRegSummaryIssues: {
-                        ...newIssues,
-                        results: [
-                            ...(oldIssues.results ?? []),
-                            ...(newIssues.results ?? []),
-                        ],
-                    },
-                });
-            },
-        });
-    }, [
-        fetchMore,
-        variables,
-        data?.assessmentRegSummaryIssues?.page,
-    ]);
-
     const getError = useCallback(
         (clientId?: string) => {
             if (!clientId) {
@@ -245,21 +149,14 @@ function SelectIssueInput(props: Props) {
 
     return (
         <div className={styles.input}>
-            <SearchSelectInput
+            <IssueSearchSelectInput
                 placeholder={placeholder}
                 name="summaryIssue"
                 value={value?.summaryIssue}
+                subDimension={subDimension}
                 onChange={handleIssueChange}
-                keySelector={keySelector}
-                labelSelector={labelSelector}
                 options={options}
                 onOptionsChange={setOptions}
-                onSearchValueChange={setSearchText}
-                searchOptions={data?.assessmentRegSummaryIssues?.results}
-                totalOptionsCount={data?.assessmentRegSummaryIssues?.totalCount ?? 0}
-                optionsPending={loading}
-                onShowDropdownChange={setOpened}
-                handleShowMoreClick={handleShowMoreClick}
                 error={getError(value?.clientId)?.summaryIssue}
             />
             <TextInput
